+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.2.0 (2016-04-08)
+
+* Added the method `nodeToString` to instances of Profile.
+* Added IDs to nodes, triples and graphs.
+
 ## 1.1.0 (2016-04-07)
 
-Added a parser and writer for Turtle strings.
+* Added a parser and writer for Turtle strings.